Agregacion de ruta propia para perfil de usuario
