+++ my-app/src/app/home/homePage/userPerfil/page.tsx
@@ -1,0 +1,5 @@
+import Terminos from '@/app/home/homePage/userPerfil/UserPerfil';
+
+export default function Page() {
+  return <Terminos />;
+}
